relex/requirements.txt
absl-py==0.7.1
astor==0.8.0
atomicwrites==1.3.0
attrs==19.1.0
blis==0.2.4
certifi==2019.3.9
chardet==3.0.4
cycler==0.10.0
cymem==2.0.2
dill==0.2.9
gast==0.2.2
google-pasta==0.1.7
grpcio==1.21.1
h5py==2.9.0
idna==2.8
importlib-metadata==0.17
joblib==0.13.2
jsonschema==2.6.0
Keras==2.2.4
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
Markdown==3.1.1
matplotlib==3.1.0
mock==3.0.5
more-itertools==7.0.0
msgpack==0.5.6
msgpack-numpy==0.4.3.2
multiprocess==0.70.7
murmurhash==1.0.2
numpy==1.16.1
pathos==0.2.3
Pillow==6.0.0
plac==0.9.6
pluggy==0.12.0
pox==0.2.5
ppft==1.6.4.9
preshed==2.0.1
protobuf==3.8.0
py==1.8.0
pyparsing==2.4.0
pytest==4.5.0
python-crfsuite==0.9.6
python-dateutil==2.8.0
PyYAML==5.1.1
requests==2.22.0
scikit-learn==0.21.2
scipy==1.3.0
six==1.12.0
sklearn-crfsuite==0.3.6
spacy==2.1.3
sphinx==1.7.5
srsly==0.0.5
tabulate==0.8.3
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
thinc==7.0.4
torch==1.1.0
torchtext==0.3.1
torchvision==0.3.0
tqdm==4.32.1
url==0.4.2
urllib3==1.25.3
wasabi==0.2.2
wcwidth==0.1.7
Werkzeug==0.15.4
wrapt==1.11.2
xmltodict==0.12.0
zipp==0.5.1
